% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_PTM_statistics.R
\name{CF}
\alias{CF}
\title{Complete CorrectFirst analysis}
\description{
Complete CorrectFirst analysis

Complete CorrectFirst analysis
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CF-new}{\code{CF$new()}}
\item \href{#method-CF-get_inputs}{\code{CF$get_inputs()}}
\item \href{#method-CF-get_results}{\code{CF$get_results()}}
\item \href{#method-CF-build}{\code{CF$build()}}
\item \href{#method-CF-as_container}{\code{CF$as_container()}}
\item \href{#method-CF-write_h5mu}{\code{CF$write_h5mu()}}
\item \href{#method-CF-clone}{\code{CF$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CF-new"></a>}}
\if{latex}{\out{\hypertarget{method-CF-new}{}}}
\subsection{Method \code{new()}}{
Compute or restore a complete CF result from paired inputs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CF$new(inputs, result = .compute_cf_stage(inputs))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inputs}}{Complete paired DEA stage.}

\item{\code{result}}{Completed computation, supplied when restoring a stage.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CF-get_inputs"></a>}}
\if{latex}{\out{\hypertarget{method-CF-get_inputs}{}}}
\subsection{Method \code{get_inputs()}}{
Return the paired inputs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CF$get_inputs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CF-get_results"></a>}}
\if{latex}{\out{\hypertarget{method-CF-get_results}{}}}
\subsection{Method \code{get_results()}}{
Return complete report and computation results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CF$get_results()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CF-build"></a>}}
\if{latex}{\out{\hypertarget{method-CF-build}{}}}
\subsection{Method \code{build()}}{
Build a complete next stage.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CF$build(Type, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Type}}{Target R6 class.}

\item{\code{...}}{Target arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CF-as_container"></a>}}
\if{latex}{\out{\hypertarget{method-CF-as_container}{}}}
\subsection{Method \code{as_container()}}{
Return a detached storage representation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CF$as_container()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CF-write_h5mu"></a>}}
\if{latex}{\out{\hypertarget{method-CF-write_h5mu}{}}}
\subsection{Method \code{write_h5mu()}}{
Write this complete stage atomically.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CF$write_h5mu(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Destination H5MU file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CF-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CF-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CF$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
