% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptm_mudata_storage.R
\name{compute_ptm_results_h5mu}
\alias{compute_ptm_results_h5mu}
\title{Compute all PTM statistics using only MuData}
\usage{
compute_ptm_results_h5mu(input_h5mu, output_h5mu)
}
\arguments{
\item{input_h5mu}{Paired-input stage.}

\item{output_h5mu}{Destination statistics stage.}
}
\value{
Complete statistics stage, invisibly.
}
\description{
Compute all PTM statistics using only MuData
}
