% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophosqua_copy_helpers.R
\name{copy_phospho_Integration}
\alias{copy_phospho_Integration}
\title{copy Markdown and runscript for FragPipe combined_protein.tsv}
\usage{
copy_phospho_Integration(workdir = getwd())
}
\arguments{
\item{workdir}{directory where to copy file - default is current working directory.}
}
\description{
copy Markdown and runscript for FragPipe combined_protein.tsv
}
