% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophosqua_copy_helpers.R
\name{copy_phosphoDEA_FragPipe_TMT}
\alias{copy_phosphoDEA_FragPipe_TMT}
\title{copy Markdown and runscript for FragPipe combined_protein.tsv}
\usage{
copy_phosphoDEA_FragPipe_TMT(workdir = getwd(), run_script = FALSE)
}
\arguments{
\item{workdir}{directory where to copy file - default is current working directory.}
}
\description{
copy Markdown and runscript for FragPipe combined_protein.tsv
}
