% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prophosqua_copy_helpers.R
\name{copy_ubiDEA_Spectronaut}
\alias{copy_ubiDEA_Spectronaut}
\title{copy Markdown and runscript for Spectronaut Ubi workflow}
\usage{
copy_ubiDEA_Spectronaut(workdir = getwd())
}
\arguments{
\item{workdir}{directory where to copy file - default is current working directory.}
}
\description{
copy Markdown and runscript for Spectronaut Ubi workflow
}
