% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FP_phosphoHelperFunctions_v3_202310.R
\name{doMSstatsLikeSiteNormalizationUsingProteinStatsOnComboObject}
\alias{doMSstatsLikeSiteNormalizationUsingProteinStatsOnComboObject}
\title{Apply MSstatsPTM like site normalization (adjustment) for the protein fold-change and pvalues}
\usage{
doMSstatsLikeSiteNormalizationUsingProteinStatsOnComboObject(mycombo)
}
\arguments{
\item{mycombo}{dataframe from PTMsite and protein prolfqua statistics}
}
\description{
Apply MSstatsPTM like site normalization (adjustment) for the protein fold-change and pvalues
}
