% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_PTM_statistics.R
\name{DPA_DPU}
\alias{DPA_DPU}
\title{Complete DPA and DPU analysis}
\description{
Complete DPA and DPU analysis

Complete DPA and DPU analysis
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DPA_DPU-new}{\code{DPA_DPU$new()}}
\item \href{#method-DPA_DPU-get_inputs}{\code{DPA_DPU$get_inputs()}}
\item \href{#method-DPA_DPU-get_results}{\code{DPA_DPU$get_results()}}
\item \href{#method-DPA_DPU-build}{\code{DPA_DPU$build()}}
\item \href{#method-DPA_DPU-as_container}{\code{DPA_DPU$as_container()}}
\item \href{#method-DPA_DPU-write_h5mu}{\code{DPA_DPU$write_h5mu()}}
\item \href{#method-DPA_DPU-clone}{\code{DPA_DPU$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DPA_DPU-new"></a>}}
\if{latex}{\out{\hypertarget{method-DPA_DPU-new}{}}}
\subsection{Method \code{new()}}{
Compute or restore DPA and moderated/unmoderated DPU.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DPA_DPU$new(inputs, result = .compute_dpa_dpu_from_pair(inputs$get_pair()))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inputs}}{Complete paired DEA stage.}

\item{\code{result}}{Completed computation, supplied when restoring a stage.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DPA_DPU-get_inputs"></a>}}
\if{latex}{\out{\hypertarget{method-DPA_DPU-get_inputs}{}}}
\subsection{Method \code{get_inputs()}}{
Return the paired inputs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DPA_DPU$get_inputs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DPA_DPU-get_results"></a>}}
\if{latex}{\out{\hypertarget{method-DPA_DPU-get_results}{}}}
\subsection{Method \code{get_results()}}{
Return completed DPA/DPU results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DPA_DPU$get_results()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DPA_DPU-build"></a>}}
\if{latex}{\out{\hypertarget{method-DPA_DPU-build}{}}}
\subsection{Method \code{build()}}{
Build a complete next stage.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DPA_DPU$build(Type, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Type}}{Target R6 class.}

\item{\code{...}}{Target arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DPA_DPU-as_container"></a>}}
\if{latex}{\out{\hypertarget{method-DPA_DPU-as_container}{}}}
\subsection{Method \code{as_container()}}{
Return a detached storage representation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DPA_DPU$as_container()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DPA_DPU-write_h5mu"></a>}}
\if{latex}{\out{\hypertarget{method-DPA_DPU-write_h5mu}{}}}
\subsection{Method \code{write_h5mu()}}{
Write this complete stage atomically.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DPA_DPU$write_h5mu(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Destination H5MU file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DPA_DPU-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DPA_DPU-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DPA_DPU$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
