% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_visualization.R
\name{enrichment_visualization}
\alias{enrichment_visualization}
\title{Enrichment Visualization Functions}
\description{
Plots shared by the PTM-SEA, Kinase GSEA and MEA sections of the enrichment
report.
}
