% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleN_C_dat.R
\docType{data}
\name{exampleN_C_dat}
\alias{exampleN_C_dat}
\title{Example dataset exampleN_C_dat}
\format{
## `exampleN_C_dat`
}
\usage{
exampleN_C_dat
}
\description{
A simple example dataset demonstrating groups and numeric values.
}
\keyword{datasets}
