% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{explode_multisites}
\alias{explode_multisites}
\title{each site goes into own row, if more then on site per multisite}
\usage{
explode_multisites(combined_site)
}
\description{
each site goes into own row, if more then on site per multisite
}
