% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea_io.R
\name{get_dea_parquet}
\alias{get_dea_parquet}
\title{Get Parquet File from DEA Directory}
\usage{
get_dea_parquet(dea_dir)
}
\arguments{
\item{dea_dir}{Path to DEA output directory}
}
\value{
Full path to the parquet file
}
\description{
Get Parquet File from DEA Directory
}
\examples{
dea_dir <- file.path(tempdir(), "DEA_example")
results <- file.path(dea_dir, "Results_WU_example")
dir.create(results, recursive = TRUE, showWarnings = FALSE)
file.create(file.path(results, "lfqdata_normalized.parquet"))

basename(get_dea_parquet(dea_dir))
}
