% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea_io.R
\name{get_dea_sample_name_column}
\alias{get_dea_sample_name_column}
\title{Get the Declared Sample Column from a DEA Directory}
\usage{
get_dea_sample_name_column(dea_dir)
}
\arguments{
\item{dea_dir}{Path to a DEA output directory}
}
\value{
The sample column name
}
\description{
Get the Declared Sample Column from a DEA Directory
}
\examples{
dea_dir <- file.path(tempdir(), "DEA_example")
results <- file.path(dea_dir, "Results_WU_example")
dir.create(results, recursive = TRUE, showWarnings = FALSE)
writeLines("sample_name: sampleName", file.path(results, "lfqdata.yaml"))

get_dea_sample_name_column(dea_dir)
}
