% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea_io.R
\name{get_dea_yaml}
\alias{get_dea_yaml}
\title{Get YAML Config from DEA Directory}
\usage{
get_dea_yaml(dea_dir)
}
\arguments{
\item{dea_dir}{Path to DEA output directory}
}
\value{
Full path to the yaml file
}
\description{
Get YAML Config from DEA Directory
}
\examples{
dea_dir <- file.path(tempdir(), "DEA_example")
results <- file.path(dea_dir, "Results_WU_example")
dir.create(results, recursive = TRUE, showWarnings = FALSE)
file.create(file.path(results, "lfqdata.yaml"))

basename(get_dea_yaml(dea_dir))
}
