% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractSequenceWindows.R
\name{get_sequence_windows}
\alias{get_sequence_windows}
\title{extract sequence window from fasta files for plotting.}
\usage{
get_sequence_windows(
  phosRes,
  fasta_file,
  rev_pattern = "rev_",
  window_size = 15
)
}
\description{
extract sequence window from fasta files for plotting.
}
