% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/FP_phosphoHelperFunctions_v3_202310_DEPRECATED.R
\name{getSequenceWindowForLogo}
\alias{getSequenceWindowForLogo}
\title{prepare longformat from FragPipe Site Centric Phospho output called STY_79.9663.tsv}
\usage{
getSequenceWindowForLogo(poi, soi, fastaObject = myFasta, seqWindowOffset = 15)
}
\arguments{
\item{poi}{protein of interest from FP-site centric output}

\item{soi}{site of interest (parsed from index in FP-site centric output)}

\item{fastaObject}{read by seqinR}

\item{seqWindowOffset}{Sequence window offset plus minus from site}
}
\value{
string of X characters with p-Site in the middle
}
\description{
prepare longformat from FragPipe Site Centric Phospho output called STY_79.9663.tsv
}
