% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FP_phosphoHelperFunctions_v3_202310.R
\name{write_DEA}
\alias{write_DEA}
\alias{render_DEA}
\title{Write differential expression analysis results}
\usage{
write_DEA(GRP2, outpath, xlsxname = "AnalysisResults")

render_DEA(
  GRP2,
  outpath,
  htmlname = "Result2Grp",
  word = FALSE,
  markdown = "_Grp2Analysis.Rmd"
)
}
\arguments{
\item{GRP2}{return value of \code{\link{make_DEA_report}}}

\item{outpath}{path to place output}

\item{xlsxname}{file name for xlsx}

\item{htmlname}{name for html file}

\item{word}{default FALSE, if true create word document.s}

\item{markdown}{which file to render}
}
\description{
Write differential expression analysis results

Render DEA analysis report
}
\concept{workflow}
