% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleN_C_dat.R
\docType{data}
\name{n_c_integrated_df}
\alias{n_c_integrated_df}
\title{Example dataset n_c_integrated_df}
\format{
## `n_c_integrated_df`
}
\usage{
n_c_integrated_df
}
\description{
A simple example dataset demonstrating groups and numeric values.
}
\keyword{datasets}
