% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNtoC.R
\name{n_to_c_plot}
\alias{n_to_c_plot}
\title{N to C plot using ggplot2}
\usage{
n_to_c_plot(
  poi_matrix_min,
  protein_name,
  prot_length,
  contrast,
  thr_a = 0.05,
  thr_b = 0.2,
  color_protein = "yellow"
)
}
\arguments{
\item{poi_matrix_min}{data.frame with phosphorylation data}

\item{protein_name}{name of protein}

\item{prot_length}{protein length}

\item{contrast}{name of contrast}

\item{thr_a}{significance threshold (strict), default 0.05}

\item{thr_b}{significance threshold (lenient), default 0.20}

\item{color_protein}{color for protein-level bar (default: "yellow")}
}
\description{
N to C plot using ggplot2
}
\examples{
data(exampleN_C_dat)
poi_matrix_min <- exampleN_C_dat
poi_matrix_min$imputation_status <- "observed"

n_to_c_plot(subset(poi_matrix_min, protein_Id == "A0A1I9LPZ1"), "A0A1I9LPZ1", 2160, "H1FC")
n_to_c_plot(subset(poi_matrix_min, protein_Id == "A0A178US29"), "A0A178US29", 806, "H1FC")
}
