% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNtoC.R
\name{n_to_c_plot_integrated}
\alias{n_to_c_plot_integrated}
\title{N to C for integrated results}
\usage{
n_to_c_plot_integrated(
  poi_matrix_min,
  protein_name,
  prot_length,
  contrast,
  thr_a = 0.05,
  thr_b = 0.2,
  color_protein = "yellow"
)
}
\arguments{
\item{POI_matrixMin}{data.frame}
}
\description{
N to C for integrated results
}
\examples{
data(n_c_integrated_df)
n_to_c_plot_integrated(n_c_integrated_df, "A0A1I9LT44", 539, "WTFC")
}
