% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNtoC.R
\name{prepare_n_to_c_data}
\alias{prepare_n_to_c_data}
\title{Prepare data for N-to-C plotting}
\usage{
prepare_n_to_c_data(poi_matrix_min, model_site = "model_site")
}
\arguments{
\item{poi_matrix_min}{data.frame with phosphorylation data}
}
\value{
data.frame with prepared data for plotting
}
\description{
Prepare data for N-to-C plotting
}
\examples{
# example code
data(exampleN_C_dat)
poi_matrix_min <- prepare_n_to_c_data(exampleN_C_dat)

}
\keyword{internal}
