% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/FP_phosphoHelperFunctions_v3_202310_DEPRECATED.R
\name{read_fasta}
\alias{read_fasta}
\title{read AA fasta file, identify decoy entries and return fw fasta object only}
\usage{
read_fasta(FastaFileName, decoy_pattern = "^REV_")
}
\arguments{
\item{FastaFileName}{filename of fasta file}

\item{decoyPattern}{pattern for decoy accessions}
}
\value{
fastaObject without decoy sequences
}
\description{
read AA fasta file, identify decoy entries and return fw fasta object only
}
