% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_enrichment.R
\name{read_ptmsigdb}
\alias{read_ptmsigdb}
\title{Read a Filtered PTMsigDB from RDS or GMT}
\usage{
read_ptmsigdb(ptmsigdb_file)
}
\arguments{
\item{ptmsigdb_file}{Path ending in `.rds` or `.gmt`.}
}
\value{
Named list of signatures.
}
\description{
Read a Filtered PTMsigDB from RDS or GMT
}
\keyword{internal}
