% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FP_phosphoHelperFunctions_v3_202310.R
\name{readDecoyFastaNreturnFwOnly}
\alias{readDecoyFastaNreturnFwOnly}
\title{read AA fasta file, identify decoy entries and return fw fasta object only}
\usage{
readDecoyFastaNreturnFwOnly(FastaFileName, decoyPattern = "REV_")
}
\arguments{
\item{FastaFileName}{filename of fasta file}

\item{decoyPattern}{pattern for decoy accessions}
}
\value{
fastaObject without decoy sequences
}
\description{
read AA fasta file, identify decoy entries and return fw fasta object only
}
