% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FP_phosphoHelperFunctions_v3_202310.R
\name{test_diff_diff}
\alias{test_diff_diff}
\title{Test if differences of differences are significant}
\usage{
test_diff_diff(
  dfA,
  dfB,
  by,
  diff = c("diff"),
  std.err = c("std.error"),
  df = c("df")
)
}
\description{
Test if differences of differences are significant
}
