% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FP_phosphoHelperFunctions_v3_202310.R
\name{tidy_FragPipe_phospho_STYfile}
\alias{tidy_FragPipe_phospho_STYfile}
\title{prepare longformat from FragPipe Site Centric Phospho output called STY_79.9663.tsv}
\usage{
tidy_FragPipe_phospho_STYfile(
  styphosphopeptideTable,
  as_list = FALSE,
  locPnames = c(" Localization Probability"),
  locProbThreshold = 0.75,
  protIDcol = "Index",
  intnames = c(" Intensity"),
  maxlfqnames = c("MaxLFQ Intensity")
)
}
\arguments{
\item{styphosphopeptideTable}{original STY_79.9663 file from FragPipe}

\item{locPnames}{columns for individual  localization site probability}

\item{locProbThreshold}{Threshold on best Localization Probability}

\item{protIDcol}{id column (index = proteinName_site)}

\item{intnames}{tag to identify intensity columns}

\item{maxlfqnames}{tag to identify MaxLFQ intensity columns}
}
\value{
data.frame in lonng  format
}
\description{
prepare longformat from FragPipe Site Centric Phospho output called STY_79.9663.tsv
}
