% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_enrichment.R
\name{unzip_extdata}
\alias{unzip_extdata}
\title{Unpack One Bundled Example File}
\usage{
unzip_extdata(zip, member)
}
\arguments{
\item{zip}{File name under `inst/extdata`.}

\item{member}{File expected inside it.}
}
\value{
Path to the unpacked file.
}
\description{
Unpack One Bundled Example File
}
\keyword{internal}
