% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FP_phosphoHelperFunctions_v3_202310.R
\name{write_phosphoDEA_all}
\alias{write_phosphoDEA_all}
\title{write and render pdfs for phospho Rmarkdown}
\usage{
write_phosphoDEA_all(grp2, name, ZIPDIR, boxplot = TRUE)
}
\arguments{
\item{grp2}{grp2 object with all contrasts and all results}

\item{name}{experiment name refound in zip and htmls}

\item{ZIPDIR}{directory name to write results in}

\item{boxplot}{boolean to either do all boxplots or not}
}
\description{
write and render pdfs for phospho Rmarkdown
}
