% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_diff_diff.R
\name{.test_diff_diff}
\alias{.test_diff_diff}
\title{Test if differences of differences are significant}
\usage{
.test_diff_diff(
  dataframe_a,
  dataframe_b,
  by,
  diff = c("diff"),
  std_err = c("std.error"),
  df = c("df"),
  suffix_a = ".site",
  suffix_b = ".protein"
)
}
\description{
Test if differences of differences are significant
}
