% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_KinaseAssignments.R
\name{KinaseAssignments}
\alias{KinaseAssignments}
\title{Complete KinaseAssignments stage}
\description{
Complete KinaseAssignments stage

Complete KinaseAssignments stage
}
\details{
Construct through `source$build(Type, analysis)` and persist with `write_h5mu(path)`.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-KinaseAssignments-new}{\code{KinaseAssignments$new()}}
\item \href{#method-KinaseAssignments-get_source}{\code{KinaseAssignments$get_source()}}
\item \href{#method-KinaseAssignments-get_statistics}{\code{KinaseAssignments$get_statistics()}}
\item \href{#method-KinaseAssignments-get_analysis}{\code{KinaseAssignments$get_analysis()}}
\item \href{#method-KinaseAssignments-get_results}{\code{KinaseAssignments$get_results()}}
\item \href{#method-KinaseAssignments-build}{\code{KinaseAssignments$build()}}
\item \href{#method-KinaseAssignments-as_container}{\code{KinaseAssignments$as_container()}}
\item \href{#method-KinaseAssignments-write_h5mu}{\code{KinaseAssignments$write_h5mu()}}
\item \href{#method-KinaseAssignments-clone}{\code{KinaseAssignments$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KinaseAssignments-new"></a>}}
\if{latex}{\out{\hypertarget{method-KinaseAssignments-new}{}}}
\subsection{Method \code{new()}}{
Construct a complete result or restore an existing one.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KinaseAssignments$new(
  source,
  analysis,
  result = .compute_kinaseassignments_stage(source, analysis)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{source}}{Required complete preceding stage.}

\item{\code{analysis}}{DPA, DPU or CF.}

\item{\code{result}}{Completed computation, supplied during restoration.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KinaseAssignments-get_source"></a>}}
\if{latex}{\out{\hypertarget{method-KinaseAssignments-get_source}{}}}
\subsection{Method \code{get_source()}}{
Return the complete preceding stage.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KinaseAssignments$get_source()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KinaseAssignments-get_statistics"></a>}}
\if{latex}{\out{\hypertarget{method-KinaseAssignments-get_statistics}{}}}
\subsection{Method \code{get_statistics()}}{
Return the complete statistics component.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KinaseAssignments$get_statistics()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KinaseAssignments-get_analysis"></a>}}
\if{latex}{\out{\hypertarget{method-KinaseAssignments-get_analysis}{}}}
\subsection{Method \code{get_analysis()}}{
Return the analysis name.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KinaseAssignments$get_analysis()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KinaseAssignments-get_results"></a>}}
\if{latex}{\out{\hypertarget{method-KinaseAssignments-get_results}{}}}
\subsection{Method \code{get_results()}}{
Return the complete result.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KinaseAssignments$get_results()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KinaseAssignments-build"></a>}}
\if{latex}{\out{\hypertarget{method-KinaseAssignments-build}{}}}
\subsection{Method \code{build()}}{
Build a complete next stage.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KinaseAssignments$build(Type, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Type}}{Target R6 class.}

\item{\code{...}}{Target arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KinaseAssignments-as_container"></a>}}
\if{latex}{\out{\hypertarget{method-KinaseAssignments-as_container}{}}}
\subsection{Method \code{as_container()}}{
Return a detached storage representation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KinaseAssignments$as_container()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KinaseAssignments-write_h5mu"></a>}}
\if{latex}{\out{\hypertarget{method-KinaseAssignments-write_h5mu}{}}}
\subsection{Method \code{write_h5mu()}}{
Write this complete stage atomically.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KinaseAssignments$write_h5mu(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Destination H5MU file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KinaseAssignments-clone"></a>}}
\if{latex}{\out{\hypertarget{method-KinaseAssignments-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KinaseAssignments$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
