% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_KinaseGSEA.R
\name{KinaseGSEA}
\alias{KinaseGSEA}
\title{Complete KinaseGSEA stage}
\description{
Complete KinaseGSEA stage

Complete KinaseGSEA stage
}
\details{
Construct through `source$build(Type, analysis)` and persist with `write_h5mu(path)`.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-KinaseGSEA-new}{\code{KinaseGSEA$new()}}
\item \href{#method-KinaseGSEA-get_source}{\code{KinaseGSEA$get_source()}}
\item \href{#method-KinaseGSEA-get_statistics}{\code{KinaseGSEA$get_statistics()}}
\item \href{#method-KinaseGSEA-get_analysis}{\code{KinaseGSEA$get_analysis()}}
\item \href{#method-KinaseGSEA-get_results}{\code{KinaseGSEA$get_results()}}
\item \href{#method-KinaseGSEA-build}{\code{KinaseGSEA$build()}}
\item \href{#method-KinaseGSEA-as_container}{\code{KinaseGSEA$as_container()}}
\item \href{#method-KinaseGSEA-write_h5mu}{\code{KinaseGSEA$write_h5mu()}}
\item \href{#method-KinaseGSEA-clone}{\code{KinaseGSEA$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KinaseGSEA-new"></a>}}
\if{latex}{\out{\hypertarget{method-KinaseGSEA-new}{}}}
\subsection{Method \code{new()}}{
Construct a complete result or restore an existing one.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KinaseGSEA$new(
  source,
  analysis,
  result = .compute_kinasegsea_stage(source, analysis)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{source}}{Required complete preceding stage.}

\item{\code{analysis}}{DPA, DPU or CF.}

\item{\code{result}}{Completed computation, supplied during restoration.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KinaseGSEA-get_source"></a>}}
\if{latex}{\out{\hypertarget{method-KinaseGSEA-get_source}{}}}
\subsection{Method \code{get_source()}}{
Return the complete preceding stage.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KinaseGSEA$get_source()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KinaseGSEA-get_statistics"></a>}}
\if{latex}{\out{\hypertarget{method-KinaseGSEA-get_statistics}{}}}
\subsection{Method \code{get_statistics()}}{
Return the complete statistics component.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KinaseGSEA$get_statistics()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KinaseGSEA-get_analysis"></a>}}
\if{latex}{\out{\hypertarget{method-KinaseGSEA-get_analysis}{}}}
\subsection{Method \code{get_analysis()}}{
Return the analysis name.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KinaseGSEA$get_analysis()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KinaseGSEA-get_results"></a>}}
\if{latex}{\out{\hypertarget{method-KinaseGSEA-get_results}{}}}
\subsection{Method \code{get_results()}}{
Return the complete result.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KinaseGSEA$get_results()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KinaseGSEA-build"></a>}}
\if{latex}{\out{\hypertarget{method-KinaseGSEA-build}{}}}
\subsection{Method \code{build()}}{
Build a complete next stage.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KinaseGSEA$build(Type, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Type}}{Target R6 class.}

\item{\code{...}}{Target arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KinaseGSEA-as_container"></a>}}
\if{latex}{\out{\hypertarget{method-KinaseGSEA-as_container}{}}}
\subsection{Method \code{as_container()}}{
Return a detached storage representation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KinaseGSEA$as_container()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KinaseGSEA-write_h5mu"></a>}}
\if{latex}{\out{\hypertarget{method-KinaseGSEA-write_h5mu}{}}}
\subsection{Method \code{write_h5mu()}}{
Write this complete stage atomically.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KinaseGSEA$write_h5mu(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Destination H5MU file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KinaseGSEA-clone"></a>}}
\if{latex}{\out{\hypertarget{method-KinaseGSEA-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KinaseGSEA$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
