% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_KinaseInputs.R
\name{KinaseInputs}
\alias{KinaseInputs}
\title{Complete KinaseInputs stage}
\description{
Complete KinaseInputs stage

Complete KinaseInputs stage
}
\details{
Construct through `source$build(Type, analysis)` and persist with `write_h5mu(path)`.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-KinaseInputs-new}{\code{KinaseInputs$new()}}
\item \href{#method-KinaseInputs-get_source}{\code{KinaseInputs$get_source()}}
\item \href{#method-KinaseInputs-get_statistics}{\code{KinaseInputs$get_statistics()}}
\item \href{#method-KinaseInputs-get_analysis}{\code{KinaseInputs$get_analysis()}}
\item \href{#method-KinaseInputs-get_results}{\code{KinaseInputs$get_results()}}
\item \href{#method-KinaseInputs-build}{\code{KinaseInputs$build()}}
\item \href{#method-KinaseInputs-as_container}{\code{KinaseInputs$as_container()}}
\item \href{#method-KinaseInputs-write_h5mu}{\code{KinaseInputs$write_h5mu()}}
\item \href{#method-KinaseInputs-clone}{\code{KinaseInputs$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KinaseInputs-new"></a>}}
\if{latex}{\out{\hypertarget{method-KinaseInputs-new}{}}}
\subsection{Method \code{new()}}{
Construct a complete result or restore an existing one.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KinaseInputs$new(
  source,
  analysis,
  result = .compute_kinaseinputs_stage(source, analysis)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{source}}{Required complete preceding stage.}

\item{\code{analysis}}{DPA, DPU or CF.}

\item{\code{result}}{Completed computation, supplied during restoration.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KinaseInputs-get_source"></a>}}
\if{latex}{\out{\hypertarget{method-KinaseInputs-get_source}{}}}
\subsection{Method \code{get_source()}}{
Return the complete preceding stage.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KinaseInputs$get_source()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KinaseInputs-get_statistics"></a>}}
\if{latex}{\out{\hypertarget{method-KinaseInputs-get_statistics}{}}}
\subsection{Method \code{get_statistics()}}{
Return the complete statistics component.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KinaseInputs$get_statistics()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KinaseInputs-get_analysis"></a>}}
\if{latex}{\out{\hypertarget{method-KinaseInputs-get_analysis}{}}}
\subsection{Method \code{get_analysis()}}{
Return the analysis name.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KinaseInputs$get_analysis()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KinaseInputs-get_results"></a>}}
\if{latex}{\out{\hypertarget{method-KinaseInputs-get_results}{}}}
\subsection{Method \code{get_results()}}{
Return the complete result.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KinaseInputs$get_results()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KinaseInputs-build"></a>}}
\if{latex}{\out{\hypertarget{method-KinaseInputs-build}{}}}
\subsection{Method \code{build()}}{
Build a complete next stage.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KinaseInputs$build(Type, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Type}}{Target R6 class.}

\item{\code{...}}{Target arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KinaseInputs-as_container"></a>}}
\if{latex}{\out{\hypertarget{method-KinaseInputs-as_container}{}}}
\subsection{Method \code{as_container()}}{
Return a detached storage representation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KinaseInputs$as_container()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KinaseInputs-write_h5mu"></a>}}
\if{latex}{\out{\hypertarget{method-KinaseInputs-write_h5mu}{}}}
\subsection{Method \code{write_h5mu()}}{
Write this complete stage atomically.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KinaseInputs$write_h5mu(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Destination H5MU file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KinaseInputs-clone"></a>}}
\if{latex}{\out{\hypertarget{method-KinaseInputs-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KinaseInputs$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
