% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_MEA.R
\name{MEA}
\alias{MEA}
\title{Complete MEA stage}
\description{
Complete MEA stage

Complete MEA stage
}
\details{
Construct through `source$build(Type, analysis)` and persist with `write_h5mu(path)`.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MEA-new}{\code{MEA$new()}}
\item \href{#method-MEA-get_source}{\code{MEA$get_source()}}
\item \href{#method-MEA-get_statistics}{\code{MEA$get_statistics()}}
\item \href{#method-MEA-get_analysis}{\code{MEA$get_analysis()}}
\item \href{#method-MEA-get_results}{\code{MEA$get_results()}}
\item \href{#method-MEA-build}{\code{MEA$build()}}
\item \href{#method-MEA-as_container}{\code{MEA$as_container()}}
\item \href{#method-MEA-write_h5mu}{\code{MEA$write_h5mu()}}
\item \href{#method-MEA-clone}{\code{MEA$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MEA-new"></a>}}
\if{latex}{\out{\hypertarget{method-MEA-new}{}}}
\subsection{Method \code{new()}}{
Construct a complete result or restore an existing one.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MEA$new(source, analysis, result = .compute_mea_stage(source, analysis))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{source}}{Required complete preceding stage.}

\item{\code{analysis}}{DPA, DPU or CF.}

\item{\code{result}}{Completed computation, supplied during restoration.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MEA-get_source"></a>}}
\if{latex}{\out{\hypertarget{method-MEA-get_source}{}}}
\subsection{Method \code{get_source()}}{
Return the complete preceding stage.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MEA$get_source()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MEA-get_statistics"></a>}}
\if{latex}{\out{\hypertarget{method-MEA-get_statistics}{}}}
\subsection{Method \code{get_statistics()}}{
Return the complete statistics component.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MEA$get_statistics()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MEA-get_analysis"></a>}}
\if{latex}{\out{\hypertarget{method-MEA-get_analysis}{}}}
\subsection{Method \code{get_analysis()}}{
Return the analysis name.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MEA$get_analysis()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MEA-get_results"></a>}}
\if{latex}{\out{\hypertarget{method-MEA-get_results}{}}}
\subsection{Method \code{get_results()}}{
Return the complete result.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MEA$get_results()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MEA-build"></a>}}
\if{latex}{\out{\hypertarget{method-MEA-build}{}}}
\subsection{Method \code{build()}}{
Build a complete next stage.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MEA$build(Type, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Type}}{Target R6 class.}

\item{\code{...}}{Target arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MEA-as_container"></a>}}
\if{latex}{\out{\hypertarget{method-MEA-as_container}{}}}
\subsection{Method \code{as_container()}}{
Return a detached storage representation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MEA$as_container()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MEA-write_h5mu"></a>}}
\if{latex}{\out{\hypertarget{method-MEA-write_h5mu}{}}}
\subsection{Method \code{write_h5mu()}}{
Write this complete stage atomically.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MEA$write_h5mu(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Destination H5MU file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MEA-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MEA-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MEA$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
