% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_MotifEnrichment.R
\name{MotifEnrichment}
\alias{MotifEnrichment}
\title{Complete MotifEnrichment stage}
\description{
Complete MotifEnrichment stage

Complete MotifEnrichment stage
}
\details{
Construct through `source$build(Type, analysis)` and persist with `write_h5mu(path)`.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MotifEnrichment-new}{\code{MotifEnrichment$new()}}
\item \href{#method-MotifEnrichment-get_source}{\code{MotifEnrichment$get_source()}}
\item \href{#method-MotifEnrichment-get_statistics}{\code{MotifEnrichment$get_statistics()}}
\item \href{#method-MotifEnrichment-get_analysis}{\code{MotifEnrichment$get_analysis()}}
\item \href{#method-MotifEnrichment-get_results}{\code{MotifEnrichment$get_results()}}
\item \href{#method-MotifEnrichment-build}{\code{MotifEnrichment$build()}}
\item \href{#method-MotifEnrichment-as_container}{\code{MotifEnrichment$as_container()}}
\item \href{#method-MotifEnrichment-write_h5mu}{\code{MotifEnrichment$write_h5mu()}}
\item \href{#method-MotifEnrichment-clone}{\code{MotifEnrichment$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MotifEnrichment-new"></a>}}
\if{latex}{\out{\hypertarget{method-MotifEnrichment-new}{}}}
\subsection{Method \code{new()}}{
Construct a complete result or restore an existing one.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MotifEnrichment$new(
  source,
  analysis,
  result = .compute_motifenrichment_stage(source, analysis)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{source}}{Required complete preceding stage.}

\item{\code{analysis}}{DPA, DPU or CF.}

\item{\code{result}}{Completed computation, supplied during restoration.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MotifEnrichment-get_source"></a>}}
\if{latex}{\out{\hypertarget{method-MotifEnrichment-get_source}{}}}
\subsection{Method \code{get_source()}}{
Return the complete preceding stage.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MotifEnrichment$get_source()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MotifEnrichment-get_statistics"></a>}}
\if{latex}{\out{\hypertarget{method-MotifEnrichment-get_statistics}{}}}
\subsection{Method \code{get_statistics()}}{
Return the complete statistics component.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MotifEnrichment$get_statistics()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MotifEnrichment-get_analysis"></a>}}
\if{latex}{\out{\hypertarget{method-MotifEnrichment-get_analysis}{}}}
\subsection{Method \code{get_analysis()}}{
Return the analysis name.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MotifEnrichment$get_analysis()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MotifEnrichment-get_results"></a>}}
\if{latex}{\out{\hypertarget{method-MotifEnrichment-get_results}{}}}
\subsection{Method \code{get_results()}}{
Return the complete result.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MotifEnrichment$get_results()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MotifEnrichment-build"></a>}}
\if{latex}{\out{\hypertarget{method-MotifEnrichment-build}{}}}
\subsection{Method \code{build()}}{
Build a complete next stage.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MotifEnrichment$build(Type, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Type}}{Target R6 class.}

\item{\code{...}}{Target arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MotifEnrichment-as_container"></a>}}
\if{latex}{\out{\hypertarget{method-MotifEnrichment-as_container}{}}}
\subsection{Method \code{as_container()}}{
Return a detached storage representation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MotifEnrichment$as_container()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MotifEnrichment-write_h5mu"></a>}}
\if{latex}{\out{\hypertarget{method-MotifEnrichment-write_h5mu}{}}}
\subsection{Method \code{write_h5mu()}}{
Write this complete stage atomically.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MotifEnrichment$write_h5mu(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Destination H5MU file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MotifEnrichment-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MotifEnrichment-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MotifEnrichment$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
