% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{n_to_c_expression}
\alias{n_to_c_expression}
\title{plot protein and PTM expression}
\usage{
n_to_c_expression(
  combined_site_prot_long,
  contrast_name,
  FDR_threshold = 0.05,
  impute_flag = "Imputed_Mean_moderated"
)
}
\description{
plot protein and PTM expression
}
