% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNtoC.R
\name{n_to_c_expression}
\alias{n_to_c_expression}
\title{Plot protein and PTM expression}
\usage{
n_to_c_expression(
  combined_site_prot_long,
  contrast_name,
  FDR_threshold = 0.05,
  fc_threshold = 0,
  impute_flag = "Imputed_Mean_moderated"
)
}
\arguments{
\item{combined_site_prot_long}{data frame with combined site and protein data}

\item{contrast_name}{name of the contrast to plot}

\item{FDR_threshold}{FDR threshold for filtering significant sites (default 0.05)}

\item{fc_threshold}{Fold change threshold for filtering (default 0)}

\item{impute_flag}{Flag for imputed values (default "Imputed_Mean_moderated")}
}
\value{
data frame with protein_Id, protein_length, contrast, data, and plot
}
\description{
Plot protein and PTM expression
}
