% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNtoC.R
\name{N_to_C_plot}
\alias{N_to_C_plot}
\title{N to C plot using ggplot2}
\usage{
N_to_C_plot(
  POI_matrixMin,
  protein_name,
  protLength,
  contrast,
  thrA = 0.05,
  thrB = 0.2
)
}
\description{
N to C plot using ggplot2
}
