% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNtoC.R
\name{N_to_C_plot}
\alias{N_to_C_plot}
\title{N to C plot using ggplot2}
\usage{
N_to_C_plot(
  POI_matrixMin,
  protein_name,
  protLength,
  contrast,
  thrA = 0.05,
  thrB = 0.2,
  color_protein = "yellow"
)
}
\arguments{
\item{POI_matrixMin}{data.frame}

\item{protein_name}{name of protein}

\item{protLength}{protein length}

\item{contrast}{name of contrast}

\item{thrA}{significance threshold small default 0.05}

\item{thrB}{significance threshold small large default 0.20}
}
\description{
N to C plot using ggplot2
}
\examples{
data(exampleN_C_dat)
N_to_C_plot(exampleN_C_dat,"A0A1I9LPZ1",2160,"H1FC")
N_to_C_plot(exampleN_C_dat_no_prot, "A0A178US29",806,"H1FC")
}
