% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNtoC.R
\name{n_to_c_plot}
\alias{n_to_c_plot}
\title{N to C plot using ggplot2}
\usage{
n_to_c_plot(
  poi_matrix_min,
  protein_name,
  prot_length,
  contrast,
  thr_a = 0.05,
  thr_b = 0.2,
  color_protein = "yellow"
)
}
\arguments{
\item{protein_name}{name of protein}

\item{contrast}{name of contrast}

\item{POI_matrixMin}{data.frame}

\item{protLength}{protein length}

\item{thrA}{significance threshold small default 0.05}

\item{thrB}{significance threshold small large default 0.20}
}
\description{
N to C plot using ggplot2
}
\examples{
data(exampleN_C_dat)
n_to_c_plot(exampleN_C_dat, "A0A1I9LPZ1", 2160, "H1FC")
n_to_c_plot(exampleN_C_dat_no_prot, "A0A178US29", 806, "H1FC")
}
