% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNtoC.R
\name{n_to_c_plot_integrated}
\alias{n_to_c_plot_integrated}
\title{N to C plot for integrated (DPU) results}
\usage{
n_to_c_plot_integrated(
  poi_matrix_min,
  protein_name,
  prot_length,
  contrast,
  thr_a = 0.05,
  thr_b = 0.2,
  color_protein = "yellow"
)
}
\arguments{
\item{poi_matrix_min}{data.frame with integrated phospho/protein data}

\item{protein_name}{name of protein}

\item{prot_length}{protein length}

\item{contrast}{name of contrast}

\item{thr_a}{significance threshold (strict), default 0.05}

\item{thr_b}{significance threshold (lenient), default 0.20}

\item{color_protein}{color for protein-level bar (default: "yellow")}
}
\description{
N to C plot for integrated (DPU) results
}
\examples{
data(n_c_integrated_df)
n_c_integrated_df$imputation_status <- "observed"
n_to_c_plot_integrated(n_c_integrated_df, "A0A1I9LT44", 539, "WTFC")
}
