% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNtoC.R
\name{N_to_C_plot_integrated}
\alias{N_to_C_plot_integrated}
\title{N to C for integrated results}
\usage{
N_to_C_plot_integrated(
  POI_matrixMin,
  protein_name,
  protLength,
  contrast,
  thrA = 0.05,
  thrB = 0.2,
  color_protein = "yellow"
)
}
\description{
N to C for integrated results
}
\examples{
data(n_c_integrated_df)
N_to_C_plot_integrated(n_c_integrated_df,"A0A1I9LT44",539,"WTFC")
}
