% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{n_to_c_usage}
\alias{n_to_c_usage}
\title{plot PTM occupation, use after adjusting for total proteome}
\usage{
n_to_c_usage(
  data_combined_diff,
  contrast_name,
  FDR_threshold = 0.05,
  fc_threshold = 0,
  impute_flag = "Imputed_Mean_moderated",
  protein_Id = "protein_Id"
)
}
\description{
plot PTM occupation, use after adjusting for total proteome
}
