% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_PTM_enrichment.R
\name{PTM_enrichment}
\alias{PTM_enrichment}
\title{A Completed PTM Enrichment Stage}
\description{
A Completed PTM Enrichment Stage

A Completed PTM Enrichment Stage
}
\details{
One step of the enrichment of one analysis, DPA, DPU or CF, computed from the
stage before it or restored from its stored result. The steps chain as
PTM_statistics -> [PTMSEA], and PTM_statistics -> [KinaseInputs] ->
[KinaseAssignments] -> [KinaseGSEA], with KinaseAssignments ->
[MotifEnrichment] -> [MEA]. KinaseAssignments and MotifEnrichment come from
the kinase-library tool and are only restored.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PTM_enrichment-new}{\code{PTM_enrichment$new()}}
\item \href{#method-PTM_enrichment-get_source}{\code{PTM_enrichment$get_source()}}
\item \href{#method-PTM_enrichment-get_statistics}{\code{PTM_enrichment$get_statistics()}}
\item \href{#method-PTM_enrichment-get_analysis}{\code{PTM_enrichment$get_analysis()}}
\item \href{#method-PTM_enrichment-get_results}{\code{PTM_enrichment$get_results()}}
\item \href{#method-PTM_enrichment-clone}{\code{PTM_enrichment$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PTM_enrichment-new"></a>}}
\if{latex}{\out{\hypertarget{method-PTM_enrichment-new}{}}}
\subsection{Method \code{new()}}{
Validate and keep a completed stage.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PTM_enrichment$new(source, analysis, result)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{source}}{The stage this one is computed from.}

\item{\code{analysis}}{DPA, DPU or CF.}

\item{\code{result}}{The completed result.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PTM_enrichment-get_source"></a>}}
\if{latex}{\out{\hypertarget{method-PTM_enrichment-get_source}{}}}
\subsection{Method \code{get_source()}}{
Return the stage this one was computed from.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PTM_enrichment$get_source()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PTM_enrichment-get_statistics"></a>}}
\if{latex}{\out{\hypertarget{method-PTM_enrichment-get_statistics}{}}}
\subsection{Method \code{get_statistics()}}{
Return the complete statistics component.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PTM_enrichment$get_statistics()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PTM_enrichment-get_analysis"></a>}}
\if{latex}{\out{\hypertarget{method-PTM_enrichment-get_analysis}{}}}
\subsection{Method \code{get_analysis()}}{
Return the analysis name.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PTM_enrichment$get_analysis()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PTM_enrichment-get_results"></a>}}
\if{latex}{\out{\hypertarget{method-PTM_enrichment-get_results}{}}}
\subsection{Method \code{get_results()}}{
Return the complete result.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PTM_enrichment$get_results()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PTM_enrichment-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PTM_enrichment-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PTM_enrichment$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
