% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_PTM_results.R
\name{PTM_results}
\alias{PTM_results}
\title{Final PTM results with every enabled enrichment completed}
\description{
Final PTM results with every enabled enrichment completed

Final PTM results with every enabled enrichment completed
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PTM_results-new}{\code{PTM_results$new()}}
\item \href{#method-PTM_results-get_statistics}{\code{PTM_results$get_statistics()}}
\item \href{#method-PTM_results-get_enrichments}{\code{PTM_results$get_enrichments()}}
\item \href{#method-PTM_results-get_tables}{\code{PTM_results$get_tables()}}
\item \href{#method-PTM_results-as_container}{\code{PTM_results$as_container()}}
\item \href{#method-PTM_results-write_h5mu}{\code{PTM_results$write_h5mu()}}
\item \href{#method-PTM_results-clone}{\code{PTM_results$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PTM_results-new"></a>}}
\if{latex}{\out{\hypertarget{method-PTM_results-new}{}}}
\subsection{Method \code{new()}}{
Assemble all enabled analyses from their completed stages.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PTM_results$new(statistics, enrichments)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{statistics}}{Complete PTM_statistics stage.}

\item{\code{enrichments}}{List of complete PTMSEA, KinaseGSEA and MEA objects.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PTM_results-get_statistics"></a>}}
\if{latex}{\out{\hypertarget{method-PTM_results-get_statistics}{}}}
\subsection{Method \code{get_statistics()}}{
Return the complete statistics component.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PTM_results$get_statistics()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PTM_results-get_enrichments"></a>}}
\if{latex}{\out{\hypertarget{method-PTM_results-get_enrichments}{}}}
\subsection{Method \code{get_enrichments()}}{
Return the collection of complete enabled enrichment objects.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PTM_results$get_enrichments()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PTM_results-get_tables"></a>}}
\if{latex}{\out{\hypertarget{method-PTM_results-get_tables}{}}}
\subsection{Method \code{get_tables()}}{
Return standard statistics and abundance tables.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PTM_results$get_tables()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PTM_results-as_container"></a>}}
\if{latex}{\out{\hypertarget{method-PTM_results-as_container}{}}}
\subsection{Method \code{as_container()}}{
Return a detached storage representation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PTM_results$as_container()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PTM_results-write_h5mu"></a>}}
\if{latex}{\out{\hypertarget{method-PTM_results-write_h5mu}{}}}
\subsection{Method \code{write_h5mu()}}{
Write the final artifact atomically.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PTM_results$write_h5mu(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Destination H5MU file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PTM_results-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PTM_results-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PTM_results$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
