% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_PTM_statistics.R
\name{PTM_statistics}
\alias{PTM_statistics}
\title{Complete PTM statistics composed from independent analyses}
\description{
Complete PTM statistics composed from independent analyses

Complete PTM statistics composed from independent analyses
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PTM_statistics-new}{\code{PTM_statistics$new()}}
\item \href{#method-PTM_statistics-get_inputs}{\code{PTM_statistics$get_inputs()}}
\item \href{#method-PTM_statistics-get_dpa_dpu}{\code{PTM_statistics$get_dpa_dpu()}}
\item \href{#method-PTM_statistics-get_cf}{\code{PTM_statistics$get_cf()}}
\item \href{#method-PTM_statistics-get_tables}{\code{PTM_statistics$get_tables()}}
\item \href{#method-PTM_statistics-get_statistics}{\code{PTM_statistics$get_statistics()}}
\item \href{#method-PTM_statistics-build}{\code{PTM_statistics$build()}}
\item \href{#method-PTM_statistics-as_container}{\code{PTM_statistics$as_container()}}
\item \href{#method-PTM_statistics-write_h5mu}{\code{PTM_statistics$write_h5mu()}}
\item \href{#method-PTM_statistics-clone}{\code{PTM_statistics$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PTM_statistics-new"></a>}}
\if{latex}{\out{\hypertarget{method-PTM_statistics-new}{}}}
\subsection{Method \code{new()}}{
Combine completed DPA/DPU and CF analyses from the same inputs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PTM_statistics$new(dpa_dpu, cf)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dpa_dpu}}{Complete DPA_DPU stage.}

\item{\code{cf}}{Complete CF stage.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PTM_statistics-get_inputs"></a>}}
\if{latex}{\out{\hypertarget{method-PTM_statistics-get_inputs}{}}}
\subsection{Method \code{get_inputs()}}{
Return the paired inputs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PTM_statistics$get_inputs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PTM_statistics-get_dpa_dpu"></a>}}
\if{latex}{\out{\hypertarget{method-PTM_statistics-get_dpa_dpu}{}}}
\subsection{Method \code{get_dpa_dpu()}}{
Return completed DPA/DPU report inputs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PTM_statistics$get_dpa_dpu()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PTM_statistics-get_cf"></a>}}
\if{latex}{\out{\hypertarget{method-PTM_statistics-get_cf}{}}}
\subsection{Method \code{get_cf()}}{
Return completed CF report inputs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PTM_statistics$get_cf()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PTM_statistics-get_tables"></a>}}
\if{latex}{\out{\hypertarget{method-PTM_statistics-get_tables}{}}}
\subsection{Method \code{get_tables()}}{
Return the six standard delivery tables in memory.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PTM_statistics$get_tables()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PTM_statistics-get_statistics"></a>}}
\if{latex}{\out{\hypertarget{method-PTM_statistics-get_statistics}{}}}
\subsection{Method \code{get_statistics()}}{
Return this complete statistics component.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PTM_statistics$get_statistics()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PTM_statistics-build"></a>}}
\if{latex}{\out{\hypertarget{method-PTM_statistics-build}{}}}
\subsection{Method \code{build()}}{
Build a complete next stage.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PTM_statistics$build(Type, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Type}}{Target R6 class.}

\item{\code{...}}{Target arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PTM_statistics-as_container"></a>}}
\if{latex}{\out{\hypertarget{method-PTM_statistics-as_container}{}}}
\subsection{Method \code{as_container()}}{
Return a detached storage representation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PTM_statistics$as_container()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PTM_statistics-write_h5mu"></a>}}
\if{latex}{\out{\hypertarget{method-PTM_statistics-write_h5mu}{}}}
\subsection{Method \code{write_h5mu()}}{
Write this complete stage atomically.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PTM_statistics$write_h5mu(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Destination H5MU file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PTM_statistics-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PTM_statistics-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PTM_statistics$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
