% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6_PTMSEA.R
\name{PTMSEA}
\alias{PTMSEA}
\title{Complete PTMSEA stage}
\description{
Complete PTMSEA stage

Complete PTMSEA stage
}
\details{
Construct through `source$build(Type, analysis)` and persist with `write_h5mu(path)`.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PTMSEA-new}{\code{PTMSEA$new()}}
\item \href{#method-PTMSEA-get_source}{\code{PTMSEA$get_source()}}
\item \href{#method-PTMSEA-get_statistics}{\code{PTMSEA$get_statistics()}}
\item \href{#method-PTMSEA-get_analysis}{\code{PTMSEA$get_analysis()}}
\item \href{#method-PTMSEA-get_results}{\code{PTMSEA$get_results()}}
\item \href{#method-PTMSEA-build}{\code{PTMSEA$build()}}
\item \href{#method-PTMSEA-as_container}{\code{PTMSEA$as_container()}}
\item \href{#method-PTMSEA-write_h5mu}{\code{PTMSEA$write_h5mu()}}
\item \href{#method-PTMSEA-clone}{\code{PTMSEA$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PTMSEA-new"></a>}}
\if{latex}{\out{\hypertarget{method-PTMSEA-new}{}}}
\subsection{Method \code{new()}}{
Construct a complete result or restore an existing one.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PTMSEA$new(source, analysis, result = .compute_ptmsea_stage(source, analysis))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{source}}{Required complete preceding stage.}

\item{\code{analysis}}{DPA, DPU or CF.}

\item{\code{result}}{Completed computation, supplied during restoration.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PTMSEA-get_source"></a>}}
\if{latex}{\out{\hypertarget{method-PTMSEA-get_source}{}}}
\subsection{Method \code{get_source()}}{
Return the complete preceding stage.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PTMSEA$get_source()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PTMSEA-get_statistics"></a>}}
\if{latex}{\out{\hypertarget{method-PTMSEA-get_statistics}{}}}
\subsection{Method \code{get_statistics()}}{
Return the complete statistics component.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PTMSEA$get_statistics()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PTMSEA-get_analysis"></a>}}
\if{latex}{\out{\hypertarget{method-PTMSEA-get_analysis}{}}}
\subsection{Method \code{get_analysis()}}{
Return the analysis name.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PTMSEA$get_analysis()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PTMSEA-get_results"></a>}}
\if{latex}{\out{\hypertarget{method-PTMSEA-get_results}{}}}
\subsection{Method \code{get_results()}}{
Return the complete result.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PTMSEA$get_results()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PTMSEA-build"></a>}}
\if{latex}{\out{\hypertarget{method-PTMSEA-build}{}}}
\subsection{Method \code{build()}}{
Build a complete next stage.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PTMSEA$build(Type, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Type}}{Target R6 class.}

\item{\code{...}}{Target arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PTMSEA-as_container"></a>}}
\if{latex}{\out{\hypertarget{method-PTMSEA-as_container}{}}}
\subsection{Method \code{as_container()}}{
Return a detached storage representation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PTMSEA$as_container()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PTMSEA-write_h5mu"></a>}}
\if{latex}{\out{\hypertarget{method-PTMSEA-write_h5mu}{}}}
\subsection{Method \code{write_h5mu()}}{
Write this complete stage atomically.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PTMSEA$write_h5mu(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Destination H5MU file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PTMSEA-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PTMSEA-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PTMSEA$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
