% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptmsea.R
\name{run_ptmsea}
\alias{run_ptmsea}
\title{Run PTM-SEA without direction}
\usage{
run_ptmsea(
  ranks_list,
  pathways,
  min_size = 3,
  max_size = 500,
  n_perm = 1000,
  pvalueCutoff = 0.1
)
}
\arguments{
\item{ranks_list}{Named list of ranked vectors from ptmsea_data_prep().}

\item{pathways}{List of pathways from fgsea::gmtPathways().}

\item{min_size}{Integer. Minimum pathway size. Default 3.}

\item{max_size}{Integer. Maximum pathway size. Default 500.}

\item{n_perm}{Integer. Number of permutations. Default 1000.}

\item{pvalueCutoff}{Numeric. Adjusted p-value cutoff. Default 0.1.}
}
\value{
Named list of clusterProfiler gseaResult objects (one per contrast).
}
\description{
Runs GSEA on PTMsigDB pathways, ignoring direction suffixes (;u/;d).
}
