% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_diff_diff.R
\name{test_diff}
\alias{test_diff}
\title{compute MSstats like test statsitics}
\usage{
test_diff(
  phos_res,
  tot_res,
  join_column = c("protein_Id", "contrast", "description", "protein_length",
    "nr_tryptic_peptides")
)
}
\description{
compute MSstats like test statsitics
}
